pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rail-interview"
version = "0.0.1"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy", "absl-py", "randomname"
]